requirements.txt
-i https://pypi.org/simple
cachetools==5.2.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
dash==2.7.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
db-dtypes==1.0.5
flask==2.2.2
google-api-core[grpc]==2.11.0 ; python_version >= '3.7'
google-auth==2.15.0
google-cloud-bigquery==3.4.1
google-cloud-core==2.3.2 ; python_version >= '3.7'
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.4.0 ; python_version >= '3.7'
googleapis-common-protos==1.57.0 ; python_version >= '3.7'
grpcio==1.51.1 ; python_version >= '3.7'
grpcio-status==1.51.1
gunicorn==20.1.0
idna==3.4 ; python_version >= '3.5'
importlib-metadata==5.1.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
numpy==1.23.5
packaging==21.3 ; python_version >= '3.6'
pandas==1.5.2
plotly==5.11.0
proto-plus==1.22.1 ; python_version >= '3.6'
protobuf==4.21.12 ; python_version >= '3.7'
pyarrow==10.0.1 ; python_version >= '3.7'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynmea2==1.18.0
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.6
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
rsa==4.9 ; python_version >= '3.6'
setuptools==65.6.3 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smbus2==0.4.2
sparkfun-pi-servo-hat==0.9.0
sparkfun-qwiic==1.1.6
sparkfun-qwiic-adxl313==0.0.7
sparkfun-qwiic-alphanumeric==0.0.1
sparkfun-qwiic-as6212==0.0.2
sparkfun-qwiic-bme280==0.9.0
sparkfun-qwiic-button==2.0.1
sparkfun-qwiic-ccs811==0.9.4
sparkfun-qwiic-dual-encoder-reader==0.0.2
sparkfun-qwiic-eeprom==0.0.1
sparkfun-qwiic-gpio==0.0.2
sparkfun-qwiic-i2c==0.9.11
sparkfun-qwiic-icm20948==0.0.1
sparkfun-qwiic-joystick==0.9.0
sparkfun-qwiic-keypad==0.9.0
sparkfun-qwiic-kx13x==1.0.0
sparkfun-qwiic-led-stick==0.0.1
sparkfun-qwiic-max3010x==0.0.2
sparkfun-qwiic-micro-oled==0.10.0
sparkfun-qwiic-oled-base==0.0.2
sparkfun-qwiic-oled-display==0.0.2
sparkfun-qwiic-pca9685==0.9.1
sparkfun-qwiic-pir==0.0.4
sparkfun-qwiic-proximity==0.9.0
sparkfun-qwiic-relay==0.0.2
sparkfun-qwiic-rfid==2.0.0
sparkfun-qwiic-scmd==0.9.1
sparkfun-qwiic-serlcd==0.0.1
sparkfun-qwiic-sgp40==0.0.4
sparkfun-qwiic-soil-moisture-sensor==0.0.2
sparkfun-qwiic-tca9548a==0.9.0
sparkfun-qwiic-titan-gps==0.1.1
sparkfun-qwiic-twist==0.9.0
sparkfun-qwiic-vl53l1x==1.0.1
sparkfun-top-phat-button==0.0.2
sparkfun-ublox-gps==1.1.4
tenacity==8.1.0 ; python_version >= '3.6'
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.2.2 ; python_version >= '3.7'
zipp==3.11.0 ; python_version >= '3.7'
